+++ src/components/ProfileForm/utils/index.tsx
@@ -1,8 +1,8 @@
 import addresses from '../fakeData/addresses.json';
 import firstNames from '../fakeData/firstNames.json';
 import lastNames from '../fakeData/lastNames.json';
-import regionsJson from '../../TmtPage/regionJsons/regions.json';
-import municipalitiesJson from '../../TmtPage/regionJsons/municipalities.json';
+import regionsJson from '../../VacanciesPage/regionJsons/regions.json';
+import municipalitiesJson from '../../VacanciesPage/regionJsons/municipalities.json';
 
 import {
   Address,
