+++ src/components/VacanciesPage/JobPostingItem.tsx
