+++ src/components/VacanciesPage/SearchFilters.tsx
