+++ src/components/VacanciesPage/types/index.ts
