+++ src/components/VacanciesPage/utils/index.ts
